+++ bellows/config/__init__.py
@@ -7,8 +7,6 @@ from zigpy.config import (  # noqa: F401 pylint: disable=unused-import
     cv_boolean,
 )
 
-from .ezsp import EZSP_SCHEMA
-
 CONF_DEVICE_BAUDRATE = "baudrate"
 CONF_EZSP_CONFIG = "ezsp_config"
 CONF_PARAM_SRC_RTG = "source_routing"
@@ -21,6 +19,6 @@ CONFIG_SCHEMA = CONFIG_SCHEMA.extend(
     {
         vol.Required(CONF_DEVICE): SCHEMA_DEVICE,
         vol.Optional(CONF_PARAM_SRC_RTG, default=False): cv_boolean,
-        vol.Optional(CONF_EZSP_CONFIG, default={}): vol.Schema(EZSP_SCHEMA),
+        vol.Optional(CONF_EZSP_CONFIG, default={}): dict,
     }
 )
