+++ bellows/ezsp/v4/types/struct.py
@@ -1,0 +1,1 @@
+from bellows.types.struct import EzspStruct
