+++ bellows/types/named.py
@@ -98,292 +98,20 @@ class Bool(basic.uint8_t, enum.Enum):
     true = 0x01  # An alias for one, used for clarity.
 
 
-class EzspConfigId(basic.uint8_t, enum.Enum):
-    # Identifies a configuration value.
-
-    # The number of packet buffers available to the stack.  When set to the
-    # special value 0xFF, the NCP will allocate all remaining configuration RAM
-    # towards packet buffers, such that the resulting count will be the largest
-    # whole number of packet buffers that can fit into the available memory.
-    CONFIG_PACKET_BUFFER_COUNT = 0x01
-    # The maximum number of router neighbors the stack can keep track of. A
-    # neighbor is a node within radio range.
-    CONFIG_NEIGHBOR_TABLE_SIZE = 0x02
-    # The maximum number of APS retried messages the stack can be transmitting
-    # at any time.
-    CONFIG_APS_UNICAST_MESSAGE_COUNT = 0x03
-    # The maximum number of non-volatile bindings supported by the stack.
-    CONFIG_BINDING_TABLE_SIZE = 0x04
-    # The maximum number of EUI64 to network address associations that the
-    # stack can maintain for the application. (Note, the total number of such
-    # address associations maintained by the NCP is the sum of the value of
-    # this setting and the value of ::CONFIG_TRUST_CENTER_ADDRESS_CACHE_SIZE).
-    CONFIG_ADDRESS_TABLE_SIZE = 0x05
-    # The maximum number of multicast groups that the device may be a member
-    # of.
-    CONFIG_MULTICAST_TABLE_SIZE = 0x06
-    # The maximum number of destinations to which a node can route messages.
-    # This includes both messages originating at this node and those relayed
-    # for others.
-    CONFIG_ROUTE_TABLE_SIZE = 0x07
-    # The number of simultaneous route discoveries that a node will support.
-    CONFIG_DISCOVERY_TABLE_SIZE = 0x08
-    # The size of the alarm broadcast buffer.
-    CONFIG_BROADCAST_ALARM_DATA_SIZE = 0x09
-    # The size of the unicast alarm buffers allocated for end device children.
-    CONFIG_UNICAST_ALARM_DATA_SIZE = 0x0A
-    # Specifies the stack profile.
-    CONFIG_STACK_PROFILE = 0x0C
-    # The security level used for security at the MAC and network layers. The
-    # supported values are 0 (no security) and 5 (payload is encrypted and a
-    # four-byte MIC is used for authentication).
-    CONFIG_SECURITY_LEVEL = 0x0D
-    # The maximum number of hops for a message.
-    CONFIG_MAX_HOPS = 0x10
-    # The maximum number of end device children that a router will support.
-    CONFIG_MAX_END_DEVICE_CHILDREN = 0x11
-    # The maximum amount of time that the MAC will hold a message for indirect
-    # transmission to a child.
-    CONFIG_INDIRECT_TRANSMISSION_TIMEOUT = 0x12
-    # The maximum amount of time that an end device child can wait between
-    # polls. If no poll is heard within this timeout, then the parent removes
-    # the end device from its tables.
-    CONFIG_END_DEVICE_POLL_TIMEOUT = 0x13
-    # The maximum amount of time that a mobile node can wait between polls. If
-    # no poll is heard within this timeout, then the parent removes the mobile
-    # node from its tables.
-    CONFIG_MOBILE_NODE_POLL_TIMEOUT = 0x14
-    # The number of child table entries reserved for use only by mobile nodes.
-    CONFIG_RESERVED_MOBILE_CHILD_ENTRIES = 0x15
-    # Enables boost power mode and/or the alternate transmitter output.
-    CONFIG_TX_POWER_MODE = 0x17
-    # 0: Allow this node to relay messages. 1: Prevent this node from relaying
-    # messages.
-    CONFIG_DISABLE_RELAY = 0x18
-    # The maximum number of EUI64 to network address associations that the
-    # Trust Center can maintain.  These address cache entries are reserved for
-    # and reused by the Trust Center when processing device join/rejoin
-    # authentications. This cache size limits the number of overlapping joins
-    # the Trust Center can process within a narrow time window (e.g. two
-    # seconds), and thus should be set to the maximum number of near
-    # simultaneous joins the Trust Center is expected to accommodate. (Note,
-    # the total number of such address associations maintained by the NCP is
-    # the sum of the value of this setting and the value of
-    # ::CONFIG_ADDRESS_TABLE_SIZE.)
-    CONFIG_TRUST_CENTER_ADDRESS_CACHE_SIZE = 0x19
-    # The size of the source route table.
-    CONFIG_SOURCE_ROUTE_TABLE_SIZE = 0x1A
-    # The units used for timing out end devices on their parents.
-    CONFIG_END_DEVICE_POLL_TIMEOUT_SHIFT = 0x1B
-    # The number of blocks of a fragmented message that can be sent in a single
-    # window.
-    CONFIG_FRAGMENT_WINDOW_SIZE = 0x1C
-    # The time the stack will wait (in milliseconds) between sending blocks of
-    # a fragmented message.
-    CONFIG_FRAGMENT_DELAY_MS = 0x1D
-    # The size of the Key Table used for storing individual link keys (if the
-    # device is a Trust Center) or Application Link Keys (if the device is a
-    # normal node).
-    CONFIG_KEY_TABLE_SIZE = 0x1E
-    # The APS ACK timeout value. The stack waits this amount of time between
-    # resends of APS retried messages.
-    CONFIG_APS_ACK_TIMEOUT = 0x1F
-    # The duration of an active scan, in the units used by the 15.4 scan
-    # parameter (((1 << duration) + 1) * 15ms). This also controls the jitter
-    # used when responding to a beacon request.
-    CONFIG_ACTIVE_SCAN_DURATION = 0x20
-    # The time the coordinator will wait (in seconds) for a second end device
-    # bind request to arrive.
-    CONFIG_END_DEVICE_BIND_TIMEOUT = 0x21
-    # The number of PAN id conflict reports that must be received by the
-    # network manager within one minute to trigger a PAN id change.
-    CONFIG_PAN_ID_CONFLICT_REPORT_THRESHOLD = 0x22
-    # The timeout value in minutes for how long the Trust Center or a normal
-    # node waits for the ZigBee Request Key to complete. On the Trust Center
-    # this controls whether or not the device buffers the request, waiting for
-    # a matching pair of ZigBee Request Key. If the value is non-zero, the
-    # Trust Center buffers and waits for that amount of time. If the value is
-    # zero, the Trust Center does not buffer the request and immediately
-    # responds to the request.  Zero is the most compliant behavior.
-    CONFIG_REQUEST_KEY_TIMEOUT = 0x24
-    # This value indicates the size of the runtime modifiable certificate
-    # table. Normally certificates are stored in MFG tokens but this table can
-    # be used to field upgrade devices with new Smart Energy certificates.
-    # This value cannot be set, it can only be queried.
-    CONFIG_CERTIFICATE_TABLE_SIZE = 0x29
-    # This is a bitmask that controls which incoming ZDO request messages are
-    # passed to the application. The bits are defined in the
-    # EmberZdoConfigurationFlags enumeration. To see if the application is
-    # required to send a ZDO response in reply to an incoming message, the
-    # application must check the APS options bitfield within the
-    # incomingMessageHandler callback to see if the
-    # APS_OPTION_ZDO_RESPONSE_REQUIRED flag is set.
-    CONFIG_APPLICATION_ZDO_FLAGS = 0x2A
-    # The maximum number of broadcasts during a single broadcast timeout
-    # period.
-    CONFIG_BROADCAST_TABLE_SIZE = 0x2B
-    # The size of the MAC filter list table.
-    CONFIG_MAC_FILTER_TABLE_SIZE = 0x2C
-    # The number of supported networks.
-    CONFIG_SUPPORTED_NETWORKS = 0x2D
-    # Whether multicasts are sent to the RxOnWhenIdle=true address (0xFFFD) or
-    # the sleepy broadcast address (0xFFFF). The RxOnWhenIdle=true address is
-    # the ZigBee compliant destination for multicasts.
-    CONFIG_SEND_MULTICASTS_TO_SLEEPY_ADDRESS = 0x2E
-    # ZLL group address initial configuration.
-    CONFIG_ZLL_GROUP_ADDRESSES = 0x2F
-    # ZLL rssi threshold initial configuration.
-    CONFIG_ZLL_RSSI_THRESHOLD = 0x30
-    # The maximum number of pairings supported by the stack. Controllers
-    # must support at least one pairing table entry while targets must
-    # support at least five.
-    CONFIG_RF4CE_PAIRING_TABLE_SIZE = 0x31
-    # The maximum number of outgoing RF4CE packets supported by the stack.
-    CONFIG_RF4CE_PENDING_OUTGOING_PACKET_TABLE_SIZE = 0x32
-    # Toggles the mtorr flow control in the stack.
-    CONFIG_MTORR_FLOW_CONTROL = 0x33
-    # Setting the retry queue size.
-    CONFIG_RETRY_QUEUE_SIZE = 0x34
-    # Setting the new broadcast entry threshold.
-    CONFIG_NEW_BROADCAST_ENTRY_THRESHOLD = 0x35
-    # The length of time, in seconds, that a trust center will store a
-    # transient link key that a device can use to join its network. A transient
-    # key is added with a call to emberAddTransientLinkKey. After the transient
-    # key is added, it will be removed once this amount of time has passed. A
-    # joining device will not be able to use that key to join until it is added
-    # again on the trust center. The default value is 300 seconds, i.e., 5
-    # minutes.
-    CONFIG_TRANSIENT_KEY_TIMEOUT_S = 0x36
-
-
-class EzspValueId(basic.uint8_t, enum.Enum):
-    # Identifies a value.
-
-    # The contents of the node data stack token.
-    VALUE_TOKEN_STACK_NODE_DATA = 0x00
-    # The types of MAC passthrough messages that the host wishes to receive.
-    VALUE_MAC_PASSTHROUGH_FLAGS = 0x01
-    # The source address used to filter legacy EmberNet messages when the
-    # MAC_PASSTHROUGH_EMBERNET_SOURCE flag is set in
-    # VALUE_MAC_PASSTHROUGH_FLAGS.
-    VALUE_EMBERNET_PASSTHROUGH_SOURCE_ADDRESS = 0x02
-    # The number of available message buffers.
-    VALUE_FREE_BUFFERS = 0x03
-    # Selects sending synchronous callbacks in ezsp-uart.
-    VALUE_UART_SYNCH_CALLBACKS = 0x04
-    # The maximum incoming transfer size for the local node.
-    VALUE_MAXIMUM_INCOMING_TRANSFER_SIZE = 0x05
-    # The maximum outgoing transfer size for the local node.
-    VALUE_MAXIMUM_OUTGOING_TRANSFER_SIZE = 0x06
-    # A boolean indicating whether stack tokens are written to persistent
-    # storage as they change.
-    VALUE_STACK_TOKEN_WRITING = 0x07
-    # A read-only value indicating whether the stack is currently performing a
-    # rejoin.
-    VALUE_STACK_IS_PERFORMING_REJOIN = 0x08
-    # A list of EmberMacFilterMatchData values.
-    VALUE_MAC_FILTER_LIST = 0x09
-    # The Ember Extended Security Bitmask.
-    VALUE_EXTENDED_SECURITY_BITMASK = 0x0A
-    # The node short ID.
-    VALUE_NODE_SHORT_ID = 0x0B
-    # The descriptor capability of the local node.
-    VALUE_DESCRIPTOR_CAPABILITY = 0x0C
-    # The stack device request sequence number of the local node.
-    VALUE_STACK_DEVICE_REQUEST_SEQUENCE_NUMBER = 0x0D
-    # Enable or disable radio hold-off.
-    VALUE_RADIO_HOLD_OFF = 0x0E
-    # The flags field associated with the endpoint data.
-    VALUE_ENDPOINT_FLAGS = 0x0F
-    # Enable/disable the Mfg security config key settings.
-    VALUE_MFG_SECURITY_CONFIG = 0x10
-    # Retrieves the version information from the stack on the NCP.
-    VALUE_VERSION_INFO = 0x11
-    # This will get/set the rejoin reason noted by the host for a subsequent
-    # call to emberFindAndRejoinNetwork(). After a call to
-    # emberFindAndRejoinNetwork() the host's rejoin reason will be set to
-    # REJOIN_REASON_NONE. The NCP will store the rejoin reason used by the
-    # call to emberFindAndRejoinNetwork()
-    VALUE_NEXT_HOST_REJOIN_REASON = 0x12
-    # This is the reason that the last rejoin took place. This value may only
-    # be retrieved, not set. The rejoin may have been initiated by the stack
-    # (NCP) or the application (host). If a host initiated a rejoin the reason
-    # will be set by default to REJOIN_DUE_TO_APP_EVENT_1. If the application
-    # wishes to denote its own rejoin reasons it can do so by calling
-    # ezspSetValue(VALUE_HOST_REJOIN_REASON, REJOIN_DUE_TO_APP_EVENT_X). X is a
-    # number corresponding to one of the app events defined. If the NCP
-    # initiated a rejoin it will record this value internally for retrieval by
-    # ezspGetValue(VALUE_REAL_REJOIN_REASON).
-    VALUE_LAST_REJOIN_REASON = 0x13
-    # The next ZigBee sequence number.
-    VALUE_NEXT_ZIGBEE_SEQUENCE_NUMBER = 0x14
-    # CCA energy detect threshold for radio.
-    VALUE_CCA_THRESHOLD = 0x15
-    # The threshold value for a counter
-    VALUE_SET_COUNTER_THRESHOLD = 0x17
-    # Resets all counters thresholds to 0xFF
-    VALUE_RESET_COUNTER_THRESHOLDS = 0x18
-    # Clears all the counters
-    VALUE_CLEAR_COUNTERS = 0x19
-    # The device RF4CE base channel
-    VALUE_RF4CE_BASE_CHANNEL = 0x1A
-    # The RF4CE device types supported by the node
-    VALUE_RF4CE_SUPPORTED_DEVICE_TYPES_LIST = 0x1B
-    # The RF4CE profiles supported by the node
-    VALUE_RF4CE_SUPPORTED_PROFILES_LIST = 0x1C
-    # Setting this byte enables R21 behavior on the NCP.
-    VALUE_ENABLE_R21_BEHAVIOR = 0x29
-    # Configure the antenna mode(0-primary,1-secondary,2- toggle on tx ack
-    # fail).
-    VALUE_ANTENNA_MODE = 0x30
-    # The GDP binding recipient parameters
-    VALUE_RF4CE_GDP_BINDING_RECIPIENT_PARAMETERS = 0x1D
-    # The GDP binding push button stimulus received pending flag
-    VALUE_RF4CE_GDP_PUSH_BUTTON_STIMULUS_RECEIVED_PENDING_FLAG = 0x1E
-    # The GDP originator proxy flag in the advanced binding options
-    VALUE_RF4CE_GDP_BINDING_PROXY_FLAG = 0x1F
-    # The GDP application specific user s join unti_VALUE_RF4CE_MSO_USER_STRING
-    # 0x21 The MSO user string
-    VALUE_RF4CE_GDP_APPLICATION_SPECIFIC_USER_STRING = 0x20
-    # The MSO user string
-    VALUE_RF4CE_MSO_USER_STRING = 0x21
-    # The MSO binding recipient parameters
-    VALUE_RF4CE_MSO_BINDING_RECIPIENT_PARAMETERS = 0x22
-    # The NWK layer security frame counter value
-    VALUE_NWK_FRAME_COUNTER = 0x23
-    # The APS layer security frame counter value
-    VALUE_APS_FRAME_COUNTER = 0x24
-    # Sets the device type to use on the next rejoin using device type
-    VALUE_RETRY_DEVICE_TYPE = 0x25
-    # The device RF4CE base channel
-    VALUE_RF4CE_BASE_CHANNEL2 = 0x26
-    # The RF4CE device types supported by the node
-    VALUE_RF4CE_SUPPORTED_DEVICE_TYPES_LIST2 = 0x27
-    # The RF4CE profiles supported by the node
-    VALUE_RF4CE_SUPPORTED_PROFILES_LIST2 = 0x28
-    # Enable or disable packet traffic arbitration.
-    VALUE_ENABLE_PTA = 0x31
-    # Set packet traffic arbitration configuration options.
-    VALUE_PTA_OPTIONS = 0x32
-    # Configure manufacturing library options(0-non-CSMA transmits,1-CSMA transmits).
-    VALUE_MFGLIB_OPTIONS = 0x33
-
-
 class EzspExtendedValueId(basic.uint8_t, enum.Enum):
-    # Identifies a value based on specified characteristics. Each set of
-    # characteristics is unique to that value and is specified during the call
-    # to get the extended value.
+    # Identifies a value based on specified characteristics. Each set of characteristics
+    # is unique to that value and is specified during the call to get the extended
+    # value.
 
     # The flags field associated with the specified endpoint.
     EXTENDED_VALUE_ENDPOINT_FLAGS = 0x00
-    # This is the reason for the node to leave the network as well as the
-    # device that told it to leave. The leave reason is the 1st byte of the
-    # value while the node ID is the 2nd and 3rd byte. If the leave was caused
-    # due to an API call rather than an over the air message, the node ID will
-    # be UNKNOWN_NODE_ID (0xFFFD).
+    # This is the reason for the node to leave the network as well as the device that
+    # told it to leave. The leave reason is the 1st byte of the value while the node
+    # ID is the 2nd and 3rd byte. If the leave was caused due to an API call rather than
+    # an over the air message, the node ID will be UNKNOWN_NODE_ID (0xFFFD).
     EXTENDED_VALUE_LAST_LEAVE_REASON = 0x01
-    # This number of bytes of overhead required in the network frame for source
-    # routing to a particular destination.
+    # This number of bytes of overhead required in the network frame for source routing
+    # to a particular destination.
     EXTENDED_VALUE_GET_SOURCE_ROUTE_OVERHEAD = 0x02
 
 
@@ -410,110 +138,8 @@ class EmberConfigTxPowerMode(basic.uint16_t, enum.Enum):
     # RF_TX_ALT_N pins.  TX_POWER_MODE_BOOST_AND_ALTERNATE 0x03 Enable both
     # boost mode and the alternate transmitter output.
     TX_POWER_MODE_ALTERNATE = 0x02
-
-
-class EzspPolicyId(basic.uint8_t, enum.Enum):
-    # Identifies a policy.
-
-    # Controls trust center behavior.
-    TRUST_CENTER_POLICY = 0x00
-    # Controls how external binding modification requests are handled.
-    BINDING_MODIFICATION_POLICY = 0x01
-    # Controls whether the Host supplies unicast replies.
-    UNICAST_REPLIES_POLICY = 0x02
-    # Controls whether pollHandler callbacks are generated.
-    POLL_HANDLER_POLICY = 0x03
-    # Controls whether the message contents are included in the
-    # messageSentHandler callback.
-    MESSAGE_CONTENTS_IN_CALLBACK_POLICY = 0x04
-    # Controls whether the Trust Center will respond to Trust Center link key
-    # requests.
-    TC_KEY_REQUEST_POLICY = 0x05
-    # Controls whether the Trust Center will respond to application link key
-    # requests.
-    APP_KEY_REQUEST_POLICY = 0x06
-    # Controls whether ZigBee packets that appear invalid are automatically
-    # dropped by the stack. A counter will be incremented when this occurs.
-    PACKET_VALIDATE_LIBRARY_POLICY = 0x07
-    # Controls whether the stack will process ZLL messages.
-    ZLL_POLICY = 0x08
-
-
-class EzspDecisionId(basic.uint8_t, enum.Enum):
-    # Identifies a policy decision.
-
-    # Send the network key in the clear to all joining and rejoining devices.
-    ALLOW_JOINS = 0x00
-    # Send the network key in the clear to all joining devices.  Rejoining
-    # devices are sent the network key encrypted with their trust center link
-    # key. The trust center and any rejoining device are assumed to share a
-    # link key, either preconfigured or obtained under a previous policy.
-    ALLOW_JOINS_REJOINS_HAVE_LINK_KEY = 0x04
-    # Send the network key encrypted with the joining or rejoining device's
-    # trust center link key. The trust center and any joining or rejoining
-    # device are assumed to share a link key, either preconfigured or obtained
-    # under a previous policy. This is the default value for the
-    # TRUST_CENTER_POLICY.
-    ALLOW_PRECONFIGURED_KEY_JOINS = 0x01
-    # Send the network key encrypted with the rejoining device's trust center
-    # link key. The trust center and any rejoining device are assumed to share
-    # a link key, either preconfigured or obtained under a previous policy. No
-    # new devices are allowed to join.
-    ALLOW_REJOINS_ONLY = 0x02
-    # Reject all unsecured join and rejoin attempts.
-    DISALLOW_ALL_JOINS_AND_REJOINS = 0x03
-    # Take no action on trust center rejoin attempts.
-    IGNORE_TRUST_CENTER_REJOINS = 0x05
-    # BINDING_MODIFICATION_POLICY default decision. Do not allow the local
-    # binding table to be changed by remote nodes.
-    DISALLOW_BINDING_MODIFICATION = 0x10
-    # BINDING_MODIFICATION_POLICY decision.  Allow remote nodes to change
-    # the local binding table.
-    ALLOW_BINDING_MODIFICATION = 0x11
-    # BINDING_MODIFICATION_POLICY decision.  Allows remote nodes to set local
-    # binding entries only if the entries correspond to endpoints defined on
-    # the device, and for output clusters bound to those endpoints.
-    CHECK_BINDING_MODIFICATIONS_ARE_VALID_ENDPOINT_CLUSTERS = 0x12
-    # UNICAST_REPLIES_POLICY default decision.  The NCP will automatically send
-    # an empty reply (containing no payload) for every unicast received.
-    HOST_WILL_NOT_SUPPLY_REPLY = 0x20
-    # UNICAST_REPLIES_POLICY decision. The NCP will only send a reply if it
-    # receives a sendReply command from the Host.
-    HOST_WILL_SUPPLY_REPLY = 0x21
-    # POLL_HANDLER_POLICY default decision. Do not inform the Host when a child
-    # polls.
-    POLL_HANDLER_IGNORE = 0x30
-    # POLL_HANDLER_POLICY decision. Generate a pollHandler callback when a
-    # child polls.
-    POLL_HANDLER_CALLBACK = 0x31
-    # MESSAGE_CONTENTS_IN_CALLBACK_POLICY default decision. Include only the
-    # message tag in the messageSentHandler callback.
-    MESSAGE_TAG_ONLY_IN_CALLBACK = 0x40
-    # MESSAGE_CONTENTS_IN_CALLBACK_POLICY decision. Include both the message
-    # tag and the message contents in the messageSentHandler callback.
-    MESSAGE_TAG_AND_CONTENTS_IN_CALLBACK = 0x41
-    # TC_KEY_REQUEST_POLICY decision. When the Trust Center receives a request
-    # for a Trust Center link key, it will be ignored.
-    DENY_TC_KEY_REQUESTS = 0x50
-    # TC_KEY_REQUEST_POLICY decision. When the Trust Center receives a request
-    # for a Trust Center link key, it will reply to it with the corresponding
-    # key.
-    ALLOW_TC_KEY_REQUESTS = 0x51
-    # TC_KEY_REQUEST_POLICY decision. When the Trust Center receives a request
-    # for a Trust Center link key, it will generate a key to send to the
-    # joiner.
-    GENERATE_NEW_TC_LINK_KEY = 0x52
-    # APP_KEY_REQUEST_POLICY decision. When the Trust Center receives a request
-    # for an application link key, it will be ignored.
-    DENY_APP_KEY_REQUESTS = 0x60
-    # APP_KEY_REQUEST_POLICY decision. When the Trust Center receives a request
-    # for an application link key, it will randomly generate a key and send it
-    # to both partners.
-    ALLOW_APP_KEY_REQUESTS = 0x61
-    # Indicates that packet validate library checks are enabled on the NCP.
-    PACKET_VALIDATE_LIBRARY_CHECKS_ENABLED = 0x62
-    # Indicates that packet validate library checks are NOT enabled on the NCP.
-    PACKET_VALIDATE_LIBRARY_CHECKS_DISABLED = 0x63
+    # Enable both boost mode and the alternate transmitter output.
+    EMBER_TX_POWER_MODE_BOOST_AND_ALTERNATE = 0x03
 
 
 class EzspMfgTokenId(basic.uint8_t, enum.Enum):
@@ -642,8 +268,7 @@ class EzspStatus(basic.uint8_t, enum.Enum):
     ERROR_NO_RESPONSE = 0x39
     # The length of the command exceeded the maximum EZSP frame length.
     ERROR_COMMAND_TOO_LONG = 0x40
-    # The UART receive queue was full causing a callback response to be
-    # dropped.
+    # The UART receive queue was full causing a callback response to be dropped.
     ERROR_QUEUE_FULL = 0x41
     # The command has been filtered out by NCP.
     ERROR_COMMAND_FILTERED = 0x42
@@ -709,6 +334,20 @@ class EzspStatus(basic.uint8_t, enum.Enum):
     ASH_ACK_RECEIVED = 0x7D
     # Sent ASH Ack
     ASH_ACK_SENT = 0x7E
+    # Received ASH Nak
+    EZSP_ASH_NAK_RECEIVED = 0x7F
+    # Sent ASH Nak
+    EZSP_ASH_NAK_SENT = 0x80
+    # Received ASH RST
+    EZSP_ASH_RST_RECEIVED = 0x81
+    # Sent ASH RST
+    EZSP_ASH_RST_SENT = 0x82
+    # ASH Status
+    EZSP_ASH_STATUS = 0x83
+    # ASH TX
+    EZSP_ASH_TX = 0x84
+    # ASH RX
+    EZSP_ASH_RX = 0x85
     # No reset or error
     NO_ERROR = 0xFF
 
@@ -889,6 +528,9 @@ class EmberStatus(basic.uint8_t, enum.Enum):
     # algorithm.) PHY_TX_CCA_FAIL 0x8D The transmit attempt failed because all
     # CCA attempts indicated that the channel was busy
     PHY_TX_BUSY = 0x8C
+    # The transmit attempt failed because all CCA attempts indicated that the channel
+    # was busy
+    EMBER_PHY_TX_CCA_FAIL = 0x8D
     # The software installed on the hardware doesn't recognize the hardware
     # radio type.
     PHY_OSCILLATOR_CHECK_FAILED = 0x8E
